Revert "Added support for multiple item matching."
This reverts commit 37fdfdbdf808ba7e544b360f33aa18fa1c75da5c.

+++ VanityMonKeyGenerator/Chaos/Internal/ByteIntegerConverter.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace Chaos.NaCl.Internal
 {
@@ -12,7 +13,7 @@ namespace Chaos.NaCl.Internal
         public static UInt32 LoadLittleEndian32(byte[] buf, int offset)
         {
             return
-                buf[offset + 0]
+                (UInt32)(buf[offset + 0])
             | (((UInt32)(buf[offset + 1])) << 8)
             | (((UInt32)(buf[offset + 2])) << 16)
             | (((UInt32)(buf[offset + 3])) << 24);
@@ -29,7 +30,7 @@ namespace Chaos.NaCl.Internal
         public static UInt64 LoadBigEndian64(byte[] buf, int offset)
         {
             return
-                buf[offset + 7]
+                (UInt64)(buf[offset + 7])
                 | (((UInt64)(buf[offset + 6])) << 8)
                 | (((UInt64)(buf[offset + 5])) << 16)
                 | (((UInt64)(buf[offset + 4])) << 24)
